Move gesture and style to hook
